+++ lib/swetrix_flutter.dart
@@ -1,13 +1,12 @@
-export 'src/context/context.dart';
+export 'src/context/swetrix_context.dart';
 export 'src/error_event.dart';
 export 'src/options.dart';
 export 'src/performance_metrics.dart';
 export 'src/request_options.dart';
 export 'src/swetrix.dart';
 export 'src/exceptions/swetrix_exception.dart';
+export 'src/exceptions/forbidden_403_not_unique.dart';
 
-export 'src/context/flutter_context_builder.dart'
-    show SwetrixFlutterContextBuilder, SwetrixFlutterEnvironment;
-export 'src/swetrix_flutter_client.dart'
-    show SwetrixFlutterClient, SharedPreferencesFactory, IpAddressCacheRule;
+export 'src/context/flutter_context_builder.dart' show SwetrixContextBuilder, SwetrixFlutterEnvironment;
+export 'src/swetrix_flutter_client.dart' show SwetrixFlutterClient, SharedPreferencesFactory, IpAddressCacheRule;
 export 'src/context/visitor/visitor_store.dart' show SwetrixVisitorStore;